Refactor Yoon nelson results into separate component
